Conflicts:
	src/app/time-tracking/components/create-dialog/create-dialog.component.ts

+++ src/app/time-tracking/components/entries/delete-entry/delete-entry.component.ts
@@ -3,6 +3,7 @@ import { MdDialogRef } from '@angular/material';
 import { TimeTrackingEntryService } from '../../../../data';
 import { Http } from '@angular/http';
 import { environment } from '../../../../../environments/environment';
+import { LoginService } from '../../../../login';
 
 @Component({
   selector: 'app-delete-entry',
@@ -18,14 +19,20 @@ export class DeleteEntryComponent implements OnInit {
 
   constructor(public dialogRef: MdDialogRef<DeleteEntryComponent>,
     public timeTrackingEntryService: TimeTrackingEntryService,
+    public loginService: LoginService,
     private http: Http) {
   }
 
   public deleteEntry() {
-    this.http.delete(this.baseUrl + "/timeentries/" + this.rowid).map(res => res.json()).subscribe(
-      () => {
-        this.dialogRef.close(true);
-      });
+    if (this.loginService.loggedIn()) {
+      this.http.delete(this.baseUrl + "/timeentries/" + this.rowid).map(res => res.json()).subscribe(
+        () => {
+          this.dialogRef.close(true);
+        });
+    } else { 
+      alert("Your token has expired. Please log in again!");
+      this.dialogRef.close(true);
+    }
   }
 
   ngOnInit() {
